fixup! fix: silence inaccurate open-files-test warning

+++ tests/libtransmission/open-files-test.cc
@@ -42,7 +42,7 @@ TEST_F(OpenFilesTest, getOpensIfNotCached)
     auto buf = std::array<char, std::size(Contents) + 1>{};
     auto bytes_read = uint64_t{};
     EXPECT_TRUE(tr_sys_file_read_at(*fd, std::data(buf), std::size(Contents), 0, &bytes_read));
-    auto const contents = std::string_view{ std::data(buf), bytes_read };
+    auto const contents = std::string_view{ std::data(buf), static_cast<size_t>(bytes_read) };
     EXPECT_EQ(Contents, contents);
 }
